Move tests in right file
